+++ packages/react-native-sdk/src/utils/push/libs/index.ts
@@ -1,3 +1,11 @@
-export * from './expo';
-export * from './vanilla';
-export * from './common';
+export * from './expoNotifications';
+export * from './expoTaskManager';
+export * from './firebaseMessaging';
+export * from './iosPushNotification';
+export * from './voipPushNotification';
+export * from './callkeep';
+
+/*
+NOTE: must keep each libs in different files
+Else on commonjs, metro doesnt resolve modules properly if one of the module is not installed
+*/
