+++ packages/react-native-sdk/src/utils/push/libs/iosPushNotification.ts
@@ -1,15 +1,7 @@
-export type { FirebaseMessagingTypes } from '@react-native-firebase/messaging';
-export type FirebaseMessagingType =
-  typeof import('@react-native-firebase/messaging').default;
 export type PushNotificationIosLib =
   typeof import('@react-native-community/push-notification-ios').default;
 
-let messaging: FirebaseMessagingType | undefined;
 let pushNotificationIosLib: PushNotificationIosLib | undefined;
-
-try {
-  messaging = require('@react-native-firebase/messaging').default;
-} catch (_e) {}
 
 try {
   pushNotificationIosLib =
@@ -23,13 +15,4 @@ export function getPushNotificationIosLib() {
     );
   }
   return pushNotificationIosLib;
-}
-
-export function getFirebaseMessagingLib() {
-  if (!messaging) {
-    throw Error(
-      'react-native-firebase library is not installed. Please see https://rnfirebase.io/messaging/usage#installation for installation instructions',
-    );
-  }
-  return messaging;
 }
